Feature/refactoring

+++ src/domain/use-cases/auth/withdraw/WithdrawUseCaseImpl.ts
@@ -2,8 +2,8 @@ import { UserRepository } from '../../../repositories/user/UserRepository';
 import { Injectable } from '@nestjs/common';
 import { WithdrawUseCase } from './WithdrawUseCase';
 import { WithDrawUseCaseParams } from './dtos/WithDrawUseCaseParams';
-import { CommonUserService } from '../../user/common/CommonUserService';
 import { WithdrawResponse } from '../response.index';
+import { UserNotFoundException } from '../../../common/exceptions/customs/UserNotFoundException';
 
 @Injectable()
 export class WithdrawUseCaseImpl implements WithdrawUseCase {
@@ -12,7 +12,7 @@ export class WithdrawUseCaseImpl implements WithdrawUseCase {
   public async execute({ userId }: WithDrawUseCaseParams): WithdrawResponse {
     const user = await this._userRepository.findOne(userId);
 
-    CommonUserService.assertUserExistence(user);
+    if (!user) throw new UserNotFoundException();
 
     await this._userRepository.delete(userId);
